Avoid including previously notified events in env
Keep a local copy of the latest event in each handler task, and ensure
it is only used once instead of potentially being reused each
max_delay.

Also add IP address to events so they can be included in the env.  Might
be nice to be able to match on it too.